dev-requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
black==26.1.0
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4        
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
ConfigArgParse==1.7.1    
coverage==7.13.3
fastapi==0.110.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.43.2
MarkupSafe==3.0.3
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pluggy==1.6.0
psutil==7.2.2
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.8.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
pytokens==0.4.1
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.0.1
requests==2.32.5
ruff==0.15.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2